Add SoftVector3, make Trackers loading use sfloat, make rad loading use sfloat, make origin centralization use sfloat

+++ nfm-world/mad/render-rewrite/MeshDamage.cs
@@ -33,9 +33,9 @@ public static class MeshDamage
                 for (var j = 0; j < mesh.Polys[i].Points.Length; j++)
                 {
                     if (UMath.Py(
-                            wheel.Position.X,
+                            (float)wheel.Position.X,
                             mesh.Polys[i].Points[j].X, // x
-                            wheel.Position.Z,
+                            (float)wheel.Position.Z,
                             mesh.Polys[i].Points[j].Z // z
                         ) < stat.Clrad)
                     {
@@ -175,9 +175,9 @@ public static class MeshDamage
                     for (var j = 0; j < mesh.Polys[i].Points.Length; j++)
                     {
                         if (UMath.Py(
-                                wheel.Position.X,
+                                (float)wheel.Position.X,
                                 mesh.Polys[i].Points[j].X, // x
-                                wheel.Position.Z,
+                                (float)wheel.Position.Z,
                                 mesh.Polys[i].Points[j].Z // z
                             ) < stat.Clrad)
                         {
@@ -315,9 +315,9 @@ public static class MeshDamage
                 for (var j = 0; j < mesh.Polys[i].Points.Length; j++)
                 {
                     if (UMath.Py(
-                            wheel.Position.X,
+                            (float)wheel.Position.X,
                             mesh.Polys[i].Points[j].X, // x
-                            wheel.Position.Z,
+                            (float)wheel.Position.Z,
                             mesh.Polys[i].Points[j].Z // z
                         ) < stat.Clrad)
                     {
